Refactor and unparallelize searching
The code is much cleaner now, especially the API. It's currently not
parallelized because rayon's `ParallelIterator` is `Sized` and can't be
turned into a trait object. The workaround seems to be to collect the
results from the transition function in the transition function itself
and have it instead return a slice or vector of vertices and letting the
search function create the `ParallelIterator` from that but I am
postponing implementing that change (it is quite simple) until more
research has been done.